Unos lokacije koriscenjem Open Layers mapa

+++ WebDeliveryService/src/beans/Location.java
@@ -4,20 +4,20 @@ import java.io.Serializable;
 
 public class Location implements Serializable{
 
-   private int latitude;
-   private int longitude;
+   private double latitude;
+   private double longitude;
    private Address address;
    
-	public int getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(int latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
-	public int getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(int longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	public Address getAddress() {
@@ -31,7 +31,4 @@ public class Location implements Serializable{
 		String ret =  address.getCity() + " "  + address.getCountry();
 		return ret.trim();
 	}
-   
-    
-
 }
